ir_proto: nec_custom: optimize PIO prog., refactor packet routines
Made nec_custom PIO program smaller with a 1-byte header storing the
length of the packet to send, can be made more generic in the future.

IR device models (just `starlight_ac`, for now) now use 1-byte packet
instead of 4-byte words and use IR protocol's routines for sending data.

+++ lib/ir_dev_models/base.h
@@ -22,7 +22,7 @@ struct base_ir_device_mod {
      * Transmits a raw packet over the air (using device-specific IR
      * modulation).
      */
-    int (*tx_raw)(void *instdata, const uint32_t *packet, uint32_t len);
+    int (*tx_raw)(void *instdata, const uint8_t *packet, uint32_t len);
 
     /**
      * Controls the on/off state of the device.
